fix: 适配场景联动
